requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
black==24.3.0
coverage==7.2.3
flake8==5.0.4
matplotlib==3.7.1
gin-config==0.5.0
pytorch-ignite==0.5.0.post2
# Note: versioning of Pytorch might be dependent on compatible CUDA version.
# Please check yourself if your Pytorch installation supports cuda (for gpu acceleration)
torch==2.3.1
lightning==2.3.0
torchmetrics==1.0.3
#pytorch-cuda==11.8
lightgbm==4.4.0
xgboost==2.1.0
imbalanced-learn==0.12.3
catboost==1.2.5
numpy==1.24.3
pandas==2.2.2
pyarrow==14.0.1
pytest==7.3.1
scikit-learn==1.5.0
tensorboard==2.12.2
tqdm==4.66.3
einops==0.6.1
hydra-core==1.3
optuna==3.6.1
optuna-integration==3.6.0
wandb==0.17.3
recipies==0.1.3
#Fixed version because of NumPy incompatibility and stale development status.
scikit-optimize-fix==0.9.1
hydra-submitit-launcher==1.2.0
pytest-runner==6.0.1
